requirements.txt
Django==6.0.1
psycopg2-binary==2.9.11
dj-database-url==3.1.0
gunicorn==24.1.1
python-dotenv==1.2.1
daphne==4.2.1
redis==7.1.0
channels_redis==4.3.0
whitenoise==6.11.0